pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "privatecloud"
version = "0.2.0"
description = "PrivateCloud: one-command installer for a Kubernetes-based private cloud (k3s + monitoring + ingress + storage)."
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
  {name="Kubenew"}
]
keywords = ["private-cloud", "kubernetes", "k3s", "installer", "terraform", "ansible", "helm"]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Intended Audience :: Developers",
  "Topic :: System :: Systems Administration",
  "Topic :: Internet :: WWW/HTTP",
]

dependencies = [
  "typer>=0.12.0",
  "rich>=13.0.0",
  "pydantic>=2.0.0",
  "pyyaml>=6.0.0",
  "jinja2>=3.1.0",
  "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.0.0",
  "build>=1.0.0",
  "twine>=4.0.0"
]

[project.scripts]
privatecloud = "privatecloud.cli:app"

[project.urls]
Homepage = "https://github.com/Kubenew/PrivateCloud"
Repository = "https://github.com/Kubenew/PrivateCloud"

[tool.setuptools.packages.find]
where = ["."]
include = ["privatecloud*"]
